#21 updated services with pipe(), renamed files and methods in stores-list and categories-list components
