+++ src/app/components/categories-list/categories-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, inject } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { DataService } from '../../service/data.service';
+import { DataService } from '../../services/store.service';
 import { CategoryComponent } from '../category/category.component';
 
 
@@ -17,7 +17,7 @@ export class CategoriesListComponent {
   service = inject(DataService)
 
   ngOnInit() {
-    this.service.getData().subscribe({
+    this.service.getStore().subscribe({
       next: response => this.categoriesList = response
     })
   }
