--- src/app/service-history.service.ts
@@ -1,13 +1,0 @@
-import { HttpClient } from '@angular/common/http';
-import { Injectable, inject } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root',
-})
-export class DataService {
-  http = inject(HttpClient);
-
-  getData() {
-    return this.http.get('https://reqres.in/api/users');
-  }
-}
